01-AI-Fundamentals/module-01-exercises/deeplearningai_prompt_engineering/requirements.txt
openai==1.93.0
python-dotenv==1.1.1
redlines==0.5.2
streamlit==1.46.1